pyproject.toml
[project]
name = "x-one-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.15.2",
    "dependency-injector>=4.46.0",
    "fastapi>=0.115.12",
    "fastapi-pagination>=0.12.34",
    "loguru>=0.7.3",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
